feat: extract micrometer support into its own module
Fixes #576. This allows getting rid of mandatory micrometer dependency
if users are not interested in having metrics.
