+++ pom.xml
@@ -78,6 +78,7 @@
         <module>operator-framework-junit5</module>
         <module>operator-framework</module>
         <module>samples</module>
+      <module>micrometer-support</module>
     </modules>
